% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{animals_data}
\alias{animals_data}
\title{Example data.frame with fake animals data}
\format{A data.frame(\code{dplyr tbl_df})}
\usage{
data(animals_data)
}
\description{
Example data.frame with fake animals data.
}
\details{
A data.frame(\code{dplyr tbl_df}) with
\itemize{
\item animal. \code{factor}, animal code
\item age \code{factor}, age category code
\item size \code{numeric}, size of the animal
}
}
\keyword{data}

