% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{animals_db_eml}
\alias{animals_db_eml}
\title{Other example eml for the fake animals data}
\format{\code{eml}}
\usage{
data(animals_db_eml)
}
\description{
Other example eml for the fake animals data
}
\details{
An \code{eml} for a dataset, not a single table.
}
\keyword{data}

