% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{animals_eml}
\alias{animals_eml}
\title{Example eml for the fake animals data}
\format{\code{eml}}
\usage{
data(animals_eml)
}
\description{
Example eml for the fake animals data
}
\details{
An \code{eml}
}
\keyword{data}
