% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_translate_factors.R
\name{eml_translate_factors}
\alias{eml_translate_factors}
\title{Translate factors into their meanings}
\usage{
eml_translate_factors(df, eml, keep = FALSE, which_table = NULL)
}
\arguments{
\item{df}{Data frame}

\item{eml}{eml metadata}

\item{keep}{Whether to keep the columns with codes. If FALSE, the meanings replace the factors previous codes.}

\item{which_table}{for which table to retrieve the factors, if there are several tables (entityName)}
}
\value{
A data.frame with translated factors
}
\description{
Translate factors into their meanings
}
\examples{
data("animals_eml")
data("animals_data")
eml_translate_factors(df = animals_data, eml = animals_eml, keep = FALSE)
eml_translate_factors(df = animals_data, eml = animals_eml, keep = TRUE)

}

